variabt bit mask, immunization fix

+++ include/agentData/agentStats.h
@@ -17,7 +17,7 @@ public:
     unsigned immunizationTimestamp = 0;
     unsigned recoveredTimestamp = std::numeric_limits<unsigned>::max();
     uint8_t immunizationCount = 0;
-    uint8_t variant = 0;
+    unsigned variant = 0;
     uint8_t infectedCount = 0;
     char worstState = 0;
     friend std::ostream& operator<<(std::ostream& os, const AgentStats& s);
